+++ react/public/css/style.css
@@ -1,0 +1,7 @@
+* {
+  font-family: "Noto Sans TC", sans-serif;
+}
+
+body {
+  background-color: rgb(185, 245, 255);
+}
