+++ react/src/components/common.module.css
@@ -1,3 +1,7 @@
 .clickable{
     cursor: pointer;
 }
+
+.label {
+  display: block;
+}
