+++ react/src/styles/LoginRegisterPage.module.css
@@ -1,11 +1,3 @@
-* {
-  font-family: "Noto Sans TC", sans-serif!important;
-}
-
-body {
-  background-color: rgb(239, 236, 204);
-}
-
 .container {
   position: absolute;
   top: 50%;
@@ -17,29 +9,29 @@ body {
   align-items: center;
 }
 
-label {
+/* .label {
   display: block;
-}
+} */
 
-input {
+.input {
   margin-bottom: 20px;
 }
 
-button {
+.button {
   margin-bottom: 20px;
   background-color: #8d99ae;
   transition: 0.2s;
 }
 
-button:hover {
+.button:hover {
   background-color: #d9d9d9;
 }
 
-span {
+.span {
   font-weight: 700;
 }
 
-span:hover {
+.span:hover {
   color: rgb(68, 68, 250);
   cursor: pointer;
 }
